Updates page layouts, adds basic padding, converts tables, etc
