+++ docs/scss/components/Endpoint.scss
@@ -3,10 +3,6 @@
         padding: 25px;
     }
 
-    &-header {
-
-    }
-
     &-title {
         padding: 10px 0;
 
@@ -40,6 +36,6 @@
     }
 
     .divider {
-     border: $border;
+        border: $border;
     }
 }
