+++ docs/scss/components/Example.scss
@@ -11,14 +11,4 @@
             cursor: pointer;
         }
     }
-
-    pre {
-        background-color: $black;
-        color: $white;
-        padding: 10px;
-
-        code {
-            @extend .font-small;
-        }
-    }
 }
