+++ docs/scss/components/Method.scss
@@ -66,10 +66,6 @@
 
     }
 
-    .TypeColumn {
-        width: 75px;
-    }
-
     .DescriptionColumn {
 
     }
