+++ docs/scss/components/Tables/cells/Columns.scss
@@ -5,3 +5,19 @@
 .EndpointColumn {
     width: 300px;
 }
+
+.ParameterColumn {
+    width: 125px;
+}
+
+.OperatorColumn {
+    width: 125px;
+}
+
+.TypeColumn {
+    width: 75px;
+}
+
+.StatusCodeColumn {
+    width: 225px;
+}
