+++ docs/scss/components/index.scss
@@ -1,5 +1,5 @@
 @import 'EndpointIndex';
 @import 'Endpoint';
 @import 'Example';
-@import 'Method';
+@import 'GettingStartedPage';
 @import 'Tables/index';
