+++ docs/scss/docs/Docs.scss
@@ -1,3 +1,11 @@
 .Docs {
-    height: 100%;
+    pre {
+        background-color: $black;
+        color: $white;
+        padding: 10px;
+
+        code {
+            @extend .font-small;
+        }
+    }
 }
