+++ docs/scss/docs/index.scss
@@ -1,3 +1,4 @@
+@improt 'defaults';
 @import 'Layout';
 @import 'VerticalNav';
 @import 'Docs';
