+++ docs/src/RoutesGenerator.js
