+++ docs/src/components/intros/index.js
@@ -1,0 +1,5 @@
+export { default as Introduction } from './Introduction';
+export { default as Access } from './Access';
+export { default as Pagination } from './Pagination';
+export { default as Filtering } from './Filtering';
+export { default as Errors } from './Errors';
