+++ docs/src/layouts/index.js
@@ -1,7 +1,2 @@
 export { default as Layout } from './Layout';
 export { default as IndexLayout } from './IndexLayout';
-export { default as Introduction } from './Introduction';
-export { default as Access } from './Access';
-export { default as Pagination } from './Pagination';
-export { default as Filtering } from './Filtering';
-export { default as Errors } from './Errors';
